requirements.txt
Django>=2.1.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
whitenoise==4.1.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
djangorestframework==3.9.1
django-filter==2.1.0
Markdown==3.0.1
django-cors-headers==2.4.0
